+++ ginga/qtw/QtHelp.py
@@ -524,7 +524,7 @@ def set_default_opengl_context():
     fmt.setVersion(req.major, req.minor)
     fmt.setProfile(QSurfaceFormat.CoreProfile)
     fmt.setDefaultFormat(fmt)
-    QtCore.QCoreApplication.setAttribute(QtCore.Qt.AA_ShareOpenGLContexts)
+    QtCore.QCoreApplication.setAttribute(QtCore.Qt.AA_ShareOpenGLContexts, True)
 
 
 def delete_widget(w):